pyproject.toml
[tool.poetry]
name = "db-insight"
version = "0.1.0"
description = ""
authors = ["Hassam Hassan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
fastapi = {extras = ["all"], version = "^0.109.0"}
uvicorn = "^0.27.0"
sentry-sdk = "^1.39.2"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic-settings = "^2.1.0"
pydantic = "^2.5.3"
alembic = "^1.13.1"
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
